requirements.txt
# Railway Production Requirements
fastapi==0.115.6
uvicorn[standard]==0.34.0
mysql-connector-python==9.4.0
PyJWT==2.8.0
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.32.5
pydantic==2.10.5